clear up explanation of microkernal
not trying to be picking about wording.  but as it currently reads "A lot of the functionality that’s in the kernel is in userspace instead" i first got the impression that code is duplicated in the kernel and userspace.  am i correct in assuming the logic either lives in one or the other?  i think "functionality that’s typically in the kernel" (or something similar) makes things clearer.