Update paging.md
If I understand this correctly, bits and bytes are mixed up in this one. After the table showing how mapping the level 1 pages actually work it says one page table needs eight bits, but above it says there are eight bytes. The former seems more logical to me, especially since you showed it with `0b10000011`.

Not a big deal, but may help to avoid confusion