Clarify make's dependency semantics
The descriptions of make's dependency semantics were a bit misleading.

Prerequisites in `make` are always file names. If a prerequisite file
exists then the dependency is satisfied, regardless of whether or not
that file is also a target of another rule.

Also, .PHONY should be applied to `default`, `build` and `run` for the
same reason that it should be applied to `clean`.