p3_table is a label: it is not all zeros
This text was confused about the difference between p3_table (a label,
i.e. an address) and the contents of the memory pointed to by p3_table.

p3_table itself is not all zeros, it is the address of a chunk of memory
that has been initialized to all zeros. Since it is 4096 byte aligned,
the bottom 12 bits of the address will be zeros, that's why you can use
the bottom bits as flags.